Refactor the TakeBestPricedOfferTo*Btc bots
The TakeBestPricedOfferTo*Btc needed to be split up so they are specific
to fiat offers, and this change adds new TakeBestPricedOfferTo*Xmr bots
just for XMR.

This simplifies the the non-bsq-swap bots;  documenting, coding,
configuring, and logging the taking of both fiat and xmr (altcoin)
offers in the same bots was getting confusing, and that's going to
turn away potential users.

This refactoring means there are six bots now, and it forced considerable
refactoring in all of them.

This commit is the bulk of the work for the PR, but there will be more
changes after everything is tested again, and comments are adjusted.